requirements.txt
# --- Core Application Dependencies ---
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
streamlit
python-dotenv
watchdog
google-generativeai
GitPython

# --- Development & Quality Tools ---
black
isort
mypy
types-requests
types-PyYAML
pytest
pytest-dotenv